Extract showDeleteDetail logic into DeleteDetailsModal file
